阅圈按热度排序

+++ models/comment.js
@@ -39,7 +39,7 @@ module.exports = function (sequelize, DataTypes) {
                 'allowNull': false
             },
             'createdAt': {
-                'type': DataTypes.INTEGER,
+                'type': DataTypes.DOUBLE,
                 'allowNull': true
             },
             'updatedAt': {
